+++ python/src/ouster/client/__init__.py
@@ -44,3 +44,9 @@ from .core import ScanSource
 from .core import Packets
 from .core import Sensor
 from .core import Scans
+from .core import first_valid_column
+from .core import last_valid_column
+from .core import first_valid_column_ts
+from .core import last_valid_column_ts
+from .core import first_valid_column_pose
+from .core import last_valid_column_pose